+++ extensions/vscode/src/api/types/events.ts
@@ -690,6 +690,9 @@ export function isPublishCreateDeploymentFailure(
 
 export interface PublishDeployContentStart extends EventStreamMessage {
   type: "publish/deployContent/start";
+  data: {
+    localId: string;
+  };
 }
 export type OnPublishDeployContentStartCallback = (
   msg: PublishDeployContentStart,
@@ -702,6 +705,9 @@ export function isPublishDeployContentStart(
 
 export interface PublishDeployContentLog extends EventStreamMessage {
   type: "publish/deployContent/log";
+  data: {
+    localId: string;
+  };
 }
 export type OnPublishDeployContentLogCallback = (
   msg: PublishDeployContentLog,
@@ -714,6 +720,9 @@ export function isPublishDeployContentLog(
 
 export interface PublishDeployContentSuccess extends EventStreamMessage {
   type: "publish/deployContent/success";
+  data: {
+    localId: string;
+  };
 }
 export type OnPublishDeployContentSuccessCallback = (
   msg: PublishDeployContentSuccess,
@@ -741,6 +750,9 @@ export function isPublishDeployContentFailure(
 
 export interface PublishUpdateContentStart extends EventStreamMessage {
   type: "publish/updateContent/start";
+  data: {
+    localId: string;
+  };
 }
 export type OnPublishUpdateContentStartCallback = (
   msg: PublishUpdateContentStart,
@@ -764,6 +776,9 @@ export function isPublishUpdateContentLog(
 
 export interface PublishUpdateContentSuccess extends EventStreamMessage {
   type: "publish/updateContent/success";
+  data: {
+    localId: string;
+  };
 }
 export type OnPublishUpdateContentSuccessCallback = (
   msg: PublishUpdateContentSuccess,