+++ extensions/vscode/src/extension.ts
@@ -20,6 +20,7 @@ import { DocumentTracker } from "./entrypointTracker";
 import { getXDGConfigProperty } from "src/utils/config";
 import { PublisherState } from "./state";
 import { PublisherAuthProvider } from "./authProvider";
+import { copySystemInfoCommand } from "src/commands";
 
 const STATE_CONTEXT = "posit.publish.state";
 
@@ -44,6 +45,13 @@ export enum SelectionCredentialMatch {
 const SELECTION_IS_PRE_CONTENT_RECORD_CONTEXT =
   "posit.publish.selection.isPreContentRecord";
 export enum SelectionIsPreContentRecord {
+  true = "true",
+  false = "false",
+}
+
+const SELECTION_IS_CONNECT_CONTENT_RECORD_CONTEXT =
+  "posit.publish.selection.isConnectContentRecord";
+export enum SelectionIsConnectContentRecord {
   true = "true",
   false = "false",
 }
@@ -75,6 +83,16 @@ export function setSelectionIsPreContentRecord(
   commands.executeCommand(
     "setContext",
     SELECTION_IS_PRE_CONTENT_RECORD_CONTEXT,
+    context,
+  );
+}
+
+export function setSelectionIsConnectContentRecord(
+  context: SelectionIsConnectContentRecord,
+) {
+  commands.executeCommand(
+    "setContext",
+    SELECTION_IS_CONNECT_CONTENT_RECORD_CONTEXT,
     context,
   );
 }
@@ -121,7 +139,7 @@ async function initializeExtension(context: ExtensionContext) {
   // Then the registration of the data providers with the VSCode framework
   projectTreeDataProvider.register();
   logsTreeDataProvider.register();
-  await homeViewProvider.register(watchers);
+  homeViewProvider.register(watchers);
 
   context.subscriptions.push(
     commands.registerCommand(Commands.InitProject, async (viewId: string) => {
@@ -134,6 +152,9 @@ async function initializeExtension(context: ExtensionContext) {
     ),
     commands.registerCommand(Commands.ShowPublishingLog, () =>
       commands.executeCommand(Commands.Logs.Focus),
+    ),
+    commands.registerCommand(Commands.HomeView.CopySystemInfo, () =>
+      copySystemInfoCommand(context),
     ),
   );
   setStateContext(PositPublishState.initialized);