+++ extensions/vscode/webviews/homeView/src/stores/home.ts
@@ -214,9 +214,10 @@ export const useHomeStore = defineStore("home", () => {
     });
   };
 
-  watch([selectedContentRecord], () =>
-    updateSelectionIsPreContentRecordState(),
-  );
+  watch([selectedContentRecord], () => {
+    updateSelectionIsPreContentRecordState();
+    updateSelectionIsConnectContentRecordState();
+  });
 
   const updateSelectionIsPreContentRecordState = () => {
     const hostConduit = useHostConduitService();
@@ -226,6 +227,19 @@ export const useHomeStore = defineStore("home", () => {
         state: isPreContentRecord(selectedContentRecord.value)
           ? "true"
           : "false",
+      },
+    });
+  };
+
+  const updateSelectionIsConnectContentRecordState = () => {
+    const hostConduit = useHostConduitService();
+    const serverType =
+      selectedContentRecord.value?.serverType || ServerType.CONNECT;
+    const productType = getProductType(serverType);
+    hostConduit.sendMsg({
+      kind: WebviewToHostMessageType.UPDATE_SELECTION_IS_CONNECT_CONTENT_RECORD,
+      content: {
+        state: isConnectProduct(productType) ? "true" : "false",
       },
     });
   };