+++ internal/publish/connect_cloud/content_request_base.go
@@ -30,7 +30,7 @@ func getCloudContentType(contentType config.ContentType) (types.ContentType, err
 	return "", fmt.Errorf("content type '%s' is not supported by Connect Cloud", contentType)
 }
 
-func (c *ServerPublisher) getContentRequestBase() (*types.ContentRequestBase, error) {
+func (c *ServerPublisher) getContentRequestBase(isFirstDeploy bool) (*types.ContentRequestBase, error) {
 
 	// Extract config details for the request
 	title := c.Config.Title
@@ -100,20 +100,26 @@ func (c *ServerPublisher) getContentRequestBase() (*types.ContentRequestBase, er
 		vanityName = cloudCfg.VanityName
 	}
 
-	return &types.ContentRequestBase{
+	revision := types.RequestRevision{
+		SourceType:    "bundle",
+		RVersion:      rVersion,
+		PythonVersion: pythonVersion,
+		ContentType:   cloudContentType,
+		AppMode:       appMode,
+		PrimaryFile:   c.Config.Entrypoint,
+	}
+	base := &types.ContentRequestBase{
 		Title:       title,
 		Description: c.Config.Description,
-		NextRevision: types.NextRevision{
-			SourceType:    "bundle",
-			RVersion:      rVersion,
-			PythonVersion: pythonVersion,
-			ContentType:   cloudContentType,
-			AppMode:       appMode,
-			PrimaryFile:   c.Config.Entrypoint,
-		},
-		Access:     access,
-		AppMode:    types.AppModeFromType(c.Config.Type),
-		Secrets:    secrets,
-		VanityName: vanityName,
-	}, nil
+		Access:      access,
+		AppMode:     types.AppModeFromType(c.Config.Type),
+		Secrets:     secrets,
+		VanityName:  vanityName,
+	}
+	if isFirstDeploy {
+		base.NextRevision = &revision
+	} else {
+		base.RevisionOverrides = &revision
+	}
+	return base, nil
 }