+++ internal/publish/r_functional_test.go
@@ -138,7 +138,7 @@ func (s *RPublishFunctionalSuite) TestGetRPackagesFunctional() {
 			}
 
 			// Actually call getRPackages
-			packageMap, err := publisher.getRPackages()
+			packageMap, err := publisher.getRPackages(false)
 
 			// Both approaches should have a valid package map which includes renv
 			s.Require().Nil(err)